pyproject.toml
[project]
name = "postbot-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.98.0",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.10",
]
